Dev troye

+++ TOMB5/game/camera.cpp
@@ -143,16 +143,16 @@ void CalculateCamera()
 			gotit = phd_atan(camera.item->pos.z_pos - item->pos.z_pos, camera.item->pos.x_pos - item->pos.x_pos) - item->pos.y_rot;
 			gotit >>= 1;
 			bounds = GetBoundsAccurate(camera.item);
-			tilt = phd_atan(shift, y - (bounds[2] + bounds[3]) / 2 - camera.item->pos.y_pos) >> 1;
+			tilt = (short)(phd_atan(shift, y - (bounds[2] + bounds[3]) / 2 - camera.item->pos.y_pos) >> 1);
 
 			if (gotit > -9100 && gotit < 9100 && tilt > -15470 && tilt < 15470)
 			{
-				change = gotit - lara.head_y_rot;
+				change = (short)(gotit - lara.head_y_rot);
 
 				if (change <= 728)
 				{
 					if (change >= -728)
-						lara.head_y_rot = gotit;
+						lara.head_y_rot = (short)gotit;
 					else
 						lara.head_y_rot -= 728;
 				}
